server/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Django==2.2.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ez-setup==0.9
Faker==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-client==0.0.1
mysql-connector-python==8.0.17
mysqlclient==1.4.6
pep8==1.7.1
Pillow==6.2.0
protobuf==3.10.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.1
wrapt==1.11.2
